+++ gitbutler-app/src/dedup.rs
@@ -1,10 +1,10 @@
-pub fn dedup(existing: &[&str], new: &str) -> String {
+pub(crate) fn dedup(existing: &[&str], new: &str) -> String {
     dedup_fmt(existing, new, " ")
 }
 
 /// Makes sure that _new_ is not in _existing_ by adding a number to it.
 /// the number is increased until the name is unique.
-pub fn dedup_fmt(existing: &[&str], new: &str, separator: &str) -> String {
+pub(crate) fn dedup_fmt(existing: &[&str], new: &str, separator: &str) -> String {
     existing
         .iter()
         .filter_map(|x| {
@@ -26,7 +26,7 @@ pub fn dedup_fmt(existing: &[&str], new: &str, separator: &str) -> String {
 }
 
 #[test]
-fn test_dedup() {
+fn tests() {
     for (existing, new, expected) in [
         (vec!["bar", "baz"], "foo", "foo"),
         (vec!["foo", "bar", "baz"], "foo", "foo 1"),
