+++ gitbutler-app/src/keys/key.rs
@@ -125,42 +125,3 @@ impl<'de> Deserialize<'de> for PublicKey {
         Self::from_str(s.as_str()).map_err(serde::de::Error::custom)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_to_from_string_private() {
-        let private_key = PrivateKey::generate();
-        let serialized = private_key.to_string();
-        let deserialized: PrivateKey = serialized.parse().unwrap();
-        assert_eq!(private_key, deserialized);
-    }
-
-    #[test]
-    fn test_to_from_string_public() {
-        let private_key = PrivateKey::generate();
-        let public_key = private_key.public_key();
-        let serialized = public_key.to_string();
-        let deserialized: PublicKey = serialized.parse().unwrap();
-        assert_eq!(public_key, deserialized);
-    }
-
-    #[test]
-    fn test_serde_private() {
-        let private_key = PrivateKey::generate();
-        let serialized = serde_json::to_string(&private_key).unwrap();
-        let deserialized: PrivateKey = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(private_key, deserialized);
-    }
-
-    #[test]
-    fn test_serde_public() {
-        let private_key = PrivateKey::generate();
-        let public_key = private_key.public_key();
-        let serialized = serde_json::to_string(&public_key).unwrap();
-        let deserialized: PublicKey = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(public_key, deserialized);
-    }
-}
