+++ gitbutler-app/src/zip.rs
@@ -163,52 +163,3 @@ fn file_hash<P: AsRef<path::Path>>(digest: &mut Sha256, path: P) -> Result<()> {
     );
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs::File;
-    use std::io::Write;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_zip_dir() {
-        let tmp_dir = tempdir().unwrap();
-        let tmp_dir_path = tmp_dir.path();
-        let file_path = tmp_dir_path.join("test.txt");
-        let mut file = File::create(file_path).unwrap();
-        file.write_all(b"test").unwrap();
-
-        let zipper_cache = tempdir().unwrap();
-        let zipper = Zipper::new(zipper_cache.path());
-        let zip_file_path = zipper.zip(tmp_dir).unwrap();
-        assert!(zip_file_path.exists());
-    }
-
-    #[test]
-    fn test_zip_file() {
-        let tmp_dir = tempdir().unwrap();
-        let tmp_dir_path = tmp_dir.path();
-        let file_path = tmp_dir_path.join("test.txt");
-        let mut file = File::create(&file_path).unwrap();
-        file.write_all(b"test").unwrap();
-
-        let zipper_cache = tempdir().unwrap();
-        let zipper = Zipper::new(zipper_cache.path());
-        zipper.zip(file_path).unwrap_err();
-    }
-
-    #[test]
-    fn test_zip_once() {
-        let tmp_dir = tempdir().unwrap();
-        let tmp_dir_path = tmp_dir.path();
-        let file_path = tmp_dir_path.join("test.txt");
-        let mut file = File::create(file_path).unwrap();
-        file.write_all(b"test").unwrap();
-
-        let zipper_cache = tempdir().unwrap();
-        let zipper = Zipper::new(zipper_cache.path());
-        assert_eq!(zipper.zip(&tmp_dir).unwrap(), zipper.zip(&tmp_dir).unwrap());
-        assert_eq!(WalkDir::new(tmp_dir).into_iter().count(), 1);
-    }
-}
