+++ gitbutler-app/tests/virtual_branches/branch/mod.rs
@@ -1,5 +1,8 @@
 use gitbutler_app::virtual_branches::Branch;
 
+mod context;
+mod file_ownership;
+mod hunk;
+mod ownership;
 mod reader;
-
 mod writer;
